UL values in XML cause NumberFormatException when > 2^31 fix #600

+++ dcm4che-core/src/main/java/org/dcm4che3/util/StringUtils.java
@@ -195,7 +195,7 @@ public class StringUtils {
 
     public static int parseIS(String s) {
         return s != null && s.length() != 0
-                ? Integer.parseInt(s.charAt(0) == '+' ? s.substring(1) : s)
+                ? (int) Long.parseLong(s.charAt(0) == '+' ? s.substring(1) : s)
                 : 0;
     }
